Refactor utility functions into utils.ts.
